EQ-309 Moved create queries to tableConfig and removed profile definitions.
