--- etl/app/server/profiles/index.js
@@ -1,7 +1,0 @@
-export * as actions from './actions.js';
-export * as assessments from './assessments.js';
-export * as assessmentUnits from './assessmentUnits.js';
-export * as assessmentUnitsMonitoringLocations from './assessmentUnitsMonitoringLocations.js';
-export * as catchmentCorrespondence from './catchmentCorrespondence.js';
-export * as sources from './sources.js';
-export * as tmdl from './tmdl.js';
